Fix crashes, link against dynamic CRT
Fix crash in free() by linking with the dynamic CRT under windows
Fix null pointer bug with a failed dynamic cast of a bound function<> to
a C-style pointer, wich was impossible. We not use a static wrapper
(HACK) to forward the call to the bound function<>. Add runtime check
before dereferencing the pointer.
Comment out OIS in the pch.