Throw "Not implemented" in GameUIManager.
We need to link with GameUIManager::setItemLevel, called from
HumanPlayer's ctor, called from PlayerManager's onLevelLoad.
But this depends directly on the GUI system, and it's not even designed
yet.
And I'm not implementing the whole GUI system just because I said I was
implementing PlayerManager and its dependencies.
Have some unconditional exception throwing in the meantime.