connected up pk_LevelTable to LobbyGM.  Allowing the lobby to read the level table
This check in connects the Level Table up to the Lobby.  So by adding
the levels in the table pk_Leveltable the lobby will allow the user to
select and play that level.  This is in contrast with the why it was
made. Where the Lobby was hard coded to show only two levels.